+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog for Flamenco Manager
 ==============================
 
+## Version 2.0.9 (released 2017-05-11)
+
+- Workers: only store host part of worker's address, and not the port number.
+
+
 ## Version 2.0.8 (released 2017-05-09)
 
 - Use mutex in scheduler to avoid race condition.